chore: split up gatekeep into multiple files.
